+++ shared/src/commonMain/kotlin/link/limecode/reddit/lite/data/model/ApiSubRedditUser.kt
@@ -5,9 +5,9 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ApiSubRedditUser(
-    val id: Int,
+    val id: Long,
     @SerialName("subreddit_id")
-    val subredditId: Int,
+    val subredditId: Long,
     @SerialName("user_id")
-    val userId: Int
+    val userId: Long
 )
